Move describe function logic into root module
